Stop trying to be overly portable
This code tried to use a Windows-specific "strcpy_s" in a code block
that was #ifdef'ed to not be compiled on Windows, using a macro to
redirect the call to a function which does almost, but not quite, the
same thing, resulting in a possible unterminated buffer vulnerability.

Stop being so silly, and use the C standard library calls instead.

Coverity CID#34600