Don't export unneeded symbols
Previously, the compiler was called without the -fvisibility
command-line parameter, which means that non-static symbols are exported
by default. This is in contrast to Windows, where the compiler by
default exports all symbols, and we need to mark those symbols that we
want to export with __declspec(dllexport).

Change CFLAGS in all Makefile.am files to use -fvisibility=hidden so
that the default for exporting symbols is changed, and add
Linux-specific definitions of the macro that resolves to the __declspec
definition on Windows so that it resolves to
__attribute__((visibility("default"))) on Linux.

This should have the effect of not exporting symbols when they're not
needed.

Still TODO: make sure we don't export a few _s symbols from
common/src/mw_util.cpp out of libbeidpkcs11.so.